move gaussian_blur_nd into functional. Add docstring
